refactor: modernize materials library flow
